### 2019-06-21 #1516 @whedon
[PRE REVIEW]: KinESim: Pre-equilibrium kinetic simulation of electrochemical reactions
**Submitting author:** @dap-biospec (<a href="http://orcid.org/0000-0003-4625-0323">Denis Proshlyakov</a>)
**Repository:** <a href="https://github.com/dap-biospec/KinESim" target ="_blank">https://github.com/dap-biospec/KinESim</a>
**Version:** v1.6.8g
