### 2019-06-21 #1516 @danielskatz
👋 @dap-biospec - your paper is missing the tags and date elements - see our [sample paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)

Once you've made changes, enter a new comment in this issue with ```@whedon generate pdf``` to rebuild the paper
