### 2019-06-21 #1516 @dap-biospec
paper.md has been updated with missing fields
