### 2019-06-21 #1516 @kyleniemeyer
@danielskatz yes, I can handle this.

However, I noticed a few major issues immediately—for one, rather than a software package, this submission seems a set of "procedures" for the proprietary software environment [Igor Pro](https://www.wavemetrics.com). Thus, it will be challenging for us to find reviewers; I'm also not totally sure if this is in scope.

There are also no tests, and the documentation is provided as a PDF & Word doc, with little commenting or in-code documentation.
