### 2019-06-21 #1516 @dap-biospec
@kyleniemeyer
Dear Kyle,

I am new to JOSS (OSS in general) and still learning the ropes. I would be happy to provide additional documentation as needed.

As for Igor Pro, it is routinely used, along with Origin and more so than Matlab, by a sizeable proportion of scientific community in chemistry, biochemistry and related disciplines. End users in those disciplines is the primary target for the software. Our submission is a complete package that can perform simulations, processing, and reporting per user needs. It allows for an extensive customization with minimal programming skills. I am write in several languages (including C++, javascript, PHP/web) and I believe that in this case this packaged form will promote open source use and development.

We submitted demo packaged Igor experiment that can used for testing. Please let me know if there is another form of testing that is preferred. Demo also illustrates adaptation for a specific type of experimental technique or conditions.

I am looking forward to working with you to make submission compliant with the accepted standards.

Best,
Denis
