### 2019-06-26 #1516 @dap-biospec
@kyleniemeyer here are several names of potential reviewers in the scientific fields related to our manuscript.  I am not affiliated with any of them.

1) Christopher Neese: physical chemistry, chemical physics, https://electroscience.osu.edu/people/neese.6

2) Tim Causgrove:  chemistry, protein kinetics and thermodynamics http://faculty.tamucc.edu/tcausgrove/Causgrove_Home.html

3) Tony Withers: geochemistry; equilibrium and dynamics; spectroscopy; - http://www.bgi.uni-bayreuth.de/?page=4&lng=en&mode=s&id=337
