### 2019-06-26 #1516 @dap-biospec
@kyleniemeyer An additional question about supplementary materials for this submission: "conventional" journals have a mechanism of submitting "for review only" materials, such as related manuscripts that did not appear in print yet. We have two submitted manuscripts that illustrate application. I think they will be valuable for reviewers to see, but cannot be uploaded to an open depository on GitHub.  I am happy to email them to you or to reviewers, or to use another mechanism if you can suggest one.
