### 2019-06-27 #1516 @kyleniemeyer
@dap-biospec we do not really have a mechanism to provide review-only materials that are not publicly available, since the entire review process happens in a GitHub issue like this (once the reviewers are finalized, this issue will be closed and the review issue will be created).

Instead, I recommend that you make your submitted manuscripts available openly via a preprint server like arXiv, ChemRxiv, PeerJ Preprints, or similar.
