### 2019-06-27 #1532 @whedon
[REVIEW]: KinESim: Pre-equilibrium kinetic simulation of electrochemical reactions
**Submitting author:** @dap-biospec (<a href="http://orcid.org/0000-0003-4625-0323">Denis Proshlyakov</a>)
**Repository:** <a href="https://github.com/dap-biospec/KinESim" target ="_blank">https://github.com/dap-biospec/KinESim</a>
**Version:** v1.6.8g

- [x] **Version:** Does the release version given match the GitHub release (v1.6.8g)?

- [x] **Version:** Does the release version given match the GitHub release (v1.6.8g)?

- [ ] **Version:** Does the release version given match the GitHub release (v1.6.8g)?
