### 2019-06-28 #1532 @FaustinCarter
I have a couple early comments about documentation:

 1. I strongly recommend converting the file KES_Manual_API.docx to a reStructured text file so that it nicely renders on github. You should also take a subset of this documentation with installation instructions, etc, and put it in your README.rst file so it shows up on the github repo landing page. Docx is a proprietary file format that is not always readable by everyone, so it isn't really a suitable format for documentation. This is really easy to do with the rstdoc tool: https://pypi.org/project/rstdoc/.

 2. As a followup to 1, you might consider using sphinx to generate documentation that the user can build, and that can optionally be hosted for free on a site like readthedocs.io. For an example of an Igor Pro related tool that does this, see: https://docs.byte-physics.de/igor-unit-testing-framework/. The rstdoc tool mentioned above will get you about 8/10 of the way there.

 3. This Igor software is strongly tied to version 8 of Igor. This is pretty new and a lot of folks may not have updated from 7 (or, like me, from 6). Do you use features that are specific to version 8, or can this be relaxed somewhat? If it can't be relaxed, the installation instructions definitely need to state upfront that it requires version 8.

 4. The existence of a pdf that overlaps with, but is not identical to, the docx file is confusing. You should have one source of truth for documentation.

OK, that's it for now. I'll have more comments as I work through testing the code. If you want help or advice for how to implement steps 1 and 2 above, I'm super happy to assist with that, but we should move that discussion to an issue in the KinESim repo.
