### 2019-06-28 #1532 @FaustinCarter
> @FaustinCarter - thank you for recommendations! We will get started on them and I will reach out to you should we run into big problems. Much of this is new to us, but luckily it takes only once to learn.

No worries! I just spent the last 10 minutes playing around with rstdoc, and it looks like you will get better conversion results if you spend a few minutes with your docx and apply styles (like heading 1, heading 2, normal text, etc) before you convert it (as opposed to just using bold, italics, and tabs to denote structure).
