### 2019-07-02 #1532 @dap-biospec
@usethedata @FaustinCarter @cfneese  @kyleniemeyer The first paper demonstrating application of this software has just been accepted. Submission files will be updated accordingly shortly.
https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859
