+++ ReadMe/ReadMe.md
@@ -1,0 +1,65 @@
+# KinESim
+Igor procedures and examples for simulation of homogeneous and heterogeneous electrochemical reactions
+
+Copyright © 2019, Denis A. Proshlyakov, dapro@chemistry.msu.edu
+
+Kin-E-Sim is a set procedures for Igor Pro designed to predict the characteristic response of an analyte in various electrochemical methods, including normal pulsed voltammetry (illustrated below), cyclic voltammetry, etc. This document is intended to provide a new user the information they need to begin using the Kin-E-Sim program. See [citation] for further details on the implemented algorithms. 
+
+Two versions of Kin-E-Sim are provided: `KES__complete.ipf` includes complete code in a single procedure file and is intended for end users.  `KES__Main.ipf` loads the same code which is split over several procedure files. It is intended for programmers who want to examine and revise the code.  Only one version can be loaded at a time. 
+
+## Installation:
+- Download `KES_complete.ipf` from the KES folder and `NPV_Illustration.ipf` and `Spectroechem.ipf` from the EChem folder. Store these files in the same folder in any location.  
+- Open Igor Pro. Select `File` -> `Open File` -> `Procedure…`
+- Locate the folder containing the necessary files. Select `KES_complete.ipf` and `Open`. 
+- A window containing the procedure will appear within the Igor workspace. Close the window. A dialog box will appear. Select `Hide`. This will compile the procedure and hide the procedure from view. Selecting `Kill` will permenantly remove the procedure from the current Igor experiment.
+- Repeat this step for `NPV_Illustrations.ipf` and `Spectroechem.ipf`.  
+- Once Kin-E-Sim is loaded and compiled, it can be accessed by selecting `Analysis` -> `Kin-E-Sim` -> `Control Panel`. The `Control Panel` will appear.  
+- Finally, select `Analysis` -> `Kin-E-Sim` -> `Create Set`. A dialog box will appear asking for a name. Any name can be entered here. Click `Continue`. All necessary waves will be generated.
+- From the `Control Panel` select the job that you just named to load the waves into the `Control Panel`.
+
+## Kin-E-Sim Demo
+A demo with pre-loaded simulations is provided in the `KES_Demo.pxp` file. A description of each pre-loaded simulation and instructions on how to perform them are provided below. Kin-E-Sim simulations are performed using the `Control Panel`. Full details of the `Control Panel` are provided in the following section. 
+
+- **Single analyte direct electrochemistry:** This simulation simulates normal pulsed voltammetry of a single analyte that undergoes direct electrochemistry on the electrode. As loaded, the analyte has a redox potential of -0.1 V. 
+		 
+- **Single analyte mediated electrochemistry:** This simulates NPV of a single analyte and mediator. The analyte is unable to perform any electron transfer on the electrode and is completely dependent on the mediator for redox transitions. As loaded, both the analyte and mediator have a redox potential of -0.1 V. The concentration of the analyte is 1 mM and the concentration of the mediator is 200 µM. 
+- **Single analyte mediator with two redox transitions:** This simulation simulates NPV of a single analyte and mediator where the mediator has two distinct redox transitions. As loaded, the analyte has a redox potential of -0.1 V and the mediator has two redox potentials at 0.0 V and -0.2 V. The concentration of the analyte is 1 mM and concentration of the mediator is 200 µM total. 
+>**Hint:** A single mediator with two distinct redox transitions is accomplished by making two separate components and connecting them using the `alias this` function. Each component must start with a concentration value greater than 0 and the sum of the two concentrations is the total concentration of the mediator.
+
+### Performing a single simulation
+
[…]
