### 2019-07-14 e5540d9
Update of documentation to conform to standards
