+++ KES_manual.md
@@ -1,0 +1,183 @@
+Return to the [readme](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/readme.md)
+
+See the [API](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/KES_API.md)
+
+## Installation:
+- Download `KES_complete.ipf` from the KES folder and `NPV_Illustration.ipf` and `Spectroechem.ipf` from the EChem folder. Store these files in the same folder in any location.  
+- Open Igor Pro. Select `File` -> `Open File` -> `Procedure…`
+- Locate the folder containing the necessary files. Select `KES_complete.ipf` and `Open`. 
+- A window containing the procedure will appear within the Igor workspace. Close the window. A dialog box will appear. Select `Hide`. This will compile the procedure and hide the procedure from view. Selecting `Kill` will permenantly remove the procedure from the current Igor experiment.
+- Repeat this step for `NPV_Illustrations.ipf` and `Spectroechem.ipf`.  
+- Once Kin-E-Sim is loaded and compiled, it can be accessed by selecting `Analysis` -> `Kin-E-Sim` -> `Control Panel`. The `Control Panel` will appear.  
+- Finally, select `Analysis` -> `Kin-E-Sim` -> `Create Set`. A dialog box will appear asking for a name. Any name can be entered here. Click `Continue`. All necessary waves will be generated.
+- From the `Control Panel` select the job that you just named to load the waves into the `Control Panel`.
+
+## Kin-E-Sim Demo
+A demo with pre-loaded simulations is provided in the `KES_Demo.pxp` file. A description of each pre-loaded simulation and instructions on how to perform them are provided below. Kin-E-Sim simulations are performed using the `Control Panel`. Full details of the `Control Panel` are provided in the following section. 
+
+- **Single analyte direct electrochemistry:** This simulation simulates normal pulsed voltammetry of a single analyte that undergoes direct electrochemistry on the electrode. As loaded, the analyte has a redox potential of -0.1 V. 
+		 
+- **Single analyte mediated electrochemistry:** This simulates NPV of a single analyte and mediator. The analyte is unable to perform any electron transfer on the electrode and is completely dependent on the mediator for redox transitions. As loaded, both the analyte and mediator have a redox potential of -0.1 V. The concentration of the analyte is 1 mM and the concentration of the mediator is 200 µM. 
+- **Single analyte mediator with two redox transitions:** This simulation simulates NPV of a single analyte and mediator where the mediator has two distinct redox transitions. As loaded, the analyte has a redox potential of -0.1 V and the mediator has two redox potentials at 0.0 V and -0.2 V. The concentration of the analyte is 1 mM and concentration of the mediator is 200 µM total. 
+>**Hint:** A single mediator with two distinct redox transitions is accomplished by making two separate components and connecting them using the `alias this` function. Each component must start with a concentration value greater than 0 and the sum of the two concentrations is the total concentration of the mediator.
+
+### Performing a single simulation
+
+- First, use the `Data Browser` (`Data` -> `Data Browser`) in Igor to select the folder with the title of the simulation you would like to run. This is done by clicking and dragging the red arrow to the desired experiment folder. Then click **[insert image]** in the top-left corner of the Kin-E-Sim `Control Panel`. The simulation should now be loaded containing all the pre-set commands and parameters.
+ 
+- To run a single simulation using the specified parameters, press the `do this sim` button in the `simulation` sub-panel. Note that the name of the simulation can be changed using the `base name` text box found in the simulation panel. Each simulation must have a unique `base name`. A new window should appear that displays the amount of time needed for the simulation to complete. Once the simulation has completed, a graph will appear containing the potential profile and concentration profiles of each component present in the simulation. 
+
[…]
