+++ README.md
@@ -1,2 +1,0 @@
-# KinESim
-Igor procedures and examples for simulation of homogeneous and heterogeneous electrochemical reactions
