### 2019-07-14 b93e7c4
Update of documentation to conform to standards
