### 2019-07-14 9467851
Update of documentation to conform to standards
