+++ KES_manual.md
@@ -1,6 +1,6 @@
-Return to the [readme](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/readme.md)
+Return to the [readme](https://github.com/dap-biospec/KinESim/blob/master/readme.md)
 
-See the [API](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/KES_API.md)
+See the [API](https://github.com/dap-biospec/KinESim/blob/master/KES_API.md)
 
 ## Installation:
 - Download `KES_complete.ipf` from the KES folder and `NPV_Illustration.ipf` and `Spectroechem.ipf` from the EChem folder. Store these files in the same folder in any location.  
@@ -57,12 +57,12 @@ Kin-E-Sim is multithreaded (MT) simulation. This is accomplished differently dep
 	- If single simulation is set to a MT mode (`CPU threads` > 1), simulations in the set will be carried out sequentially. 
 
 ## Control Panel
-![alt text](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/Figures/ControlPanel.png)
+![alt text](https://github.com/dap-biospec/KinESim/blob/master/Figures/ControlPanel.png)
 The `Control Panel` used to perform Kin-E-Sim simulations is shown above. Each sub-panel and its functions are described below:
 ### Top-left corner
-![alt text](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/Figures/TLC.png)
-- ![alt text](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/Figures/TLC_i.png): Shows the current version of Kin-E-Sim kinetic simulator
-- ![alt text](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/Figures/TLC_refresh.png): Reloads selected job
+![alt text](https://github.com/dap-biospec/KinESim/blob/master/Figures/TLC.png)
+- ![alt text](https://github.com/dap-biospec/KinESim/blob/master/Figures/TLC_i.png): Shows the current version of Kin-E-Sim kinetic simulator
+- ![alt text](https://github.com/dap-biospec/KinESim/blob/master/Figures/TLC_refresh.png): Reloads selected job
 - `job`: Allows for the selection of a user-defined job set that specifies all values, methods, and waves for the simulation. Job sets are stored in text waves. Multiple job waves can be made to easily switch between experiments. 
 - `sets=>`: Defines parameters of single simulation or simulation sets.
 - `ctrl. table`: Creates a table showing the names and contents of all parameter waves for the selected job. Parameters can be edited from this table. Press   to update changes in the Control Panel.
@@ -70,7 +70,7 @@ The `Control Panel` used to perform Kin-E-Sim simulations is shown above. Each s
 - `flags`: miscellaneous flags
 - `esim =>`: Defines the configuration of the electrochemical system and parameters of the numerical integrator. 
 ### Simulation
-![alt text](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/Figures/simulation_sp.png)
+![alt text](https://github.com/dap-biospec/KinESim/blob/master/Figures/simulation_sp.png)
 - `prep`: function that generates the potential waveform or performs other pre-simulation steps
 - `process`: function that performs step-wise simulation and integration
 - `plot`: function for post-processing of a single simulation, including preparation of a report graphs.
@@ -86,7 +86,7 @@ The `Control Panel` used to perform Kin-E-Sim simulations is shown above. Each s
 ### Integrator
 Parameters for Runge-Kutta integration
 
-![alt text](https://github.com/dap-biospec/KinESim/blob/master/KES_manual/Figures/integrator_sp.png)
+![alt text](https://github.com/dap-biospec/KinESim/blob/master/Figures/integrator_sp.png)
[…]
