### 2019-07-14 c26e6d3
Update of documentation to conform to standards
