+++ paper.bib
@@ -1,9 +1,9 @@
 @article{John:2019a,
-	Adsnote = {Placeholder for paper under re-review in Analytical Chemistry},
 	Author = {{John}, C.~W. and {Proshlyakov}, D.~A.},
 	Journal = {Analytical Chemistry},
 	Title = {{FTIR Spectro-Voltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures}},
-	Year = 2019
+	Year = 2019,
+	adsurl = {https://pubs.acs.org/doi/full/10.1021/acs.analchem.9b00859}
 }
 
 @article{John:2019b,
@@ -29,6 +29,6 @@
 	Month = feb,
 	Title = {{Molecular Computing for Markov Chains}},
 	Year = 2018,
-	adsurl = {https://arxiv.org/abs/1802.05170},
+	adsurl = {https://arxiv.org/abs/1802.05170}
 }	
 
