### 2019-07-14 #1532 @dap-biospec
@usethedata @FaustinCarter @cfneese @kyleniemeyer We updated documentation to comply with standards per @FaustinCarter recommendations. We are looking for feedback and recommendations for further improvements..
