### 2019-07-16 #1532 @kyleniemeyer
Hi @usethedata @FaustinCarter and @cfneese, just wanted to check in on your reviews.
