### 2019-07-17 #1532 @FaustinCarter
I've had a chance to look over the new documentation and it is really great! I have some further comments regarding installation and package organization:

The existence of the `KES_complete.ipf` file is confusing and provides a strong vector for unintended future version conflicts resulting from a developer updating a function in one of the sub-files and then forgetting to copy it to the right section of `KES_complete.ipf`. Instead, I think that a better solution is to create a `KinESim_Procedures` directory and a new procedure called `KinESim.ipf`. The new directory structure would be (note that `KES_complete.ipf` has been deleted):

```
KinESim_Procedures
|   KinESim.ipf
|
+---EChem
|       NPV_Illustrations.ipf
|       Spectroechem.ipf
|
\---KES
        KES_Common.ipf
        KES_Data.ipf
        KES_Panel.ipf
        KES_RKInt.ipf
        KES_Sets.ipf
        KES_Sim_Parallel.ipf
        KES_Sim_Sequential.ipf
        KES_Templates.ipf
        KES_Main.ipf
```

and the contents of `KinESim.ipf` would be:

```
#pragma rtGlobals=3
#pragma IgorVersion=8.0

#include  ":KES:KES_Main"
#include  ":EChem:NPV_Illustrations"
#include  ":EChem:Spectroechem"
```

Then the installation instructions could be simplified to: *"Clone the git repo and open the KinESim.ipf in IgorPro"*. Now all your users (end or developer) are using the same set of files, and folks can operate directly out of the cloned git repository without needing to move a bunch of files around every time you update your code.
