### 2019-07-17 #1532 @FaustinCarter
Regarding documentation: This is totally unnecessary in the sense that I'm not going to hold the review up over it, but it is good practice: I'd suggest putting all your documentation in a Docs folder in the repo (i.e. the .md files that aren't readme, and the Figures folder). This helps keep the repo clean and organized.
