### 2019-07-17 #1532 @kyleniemeyer
@FaustinCarter sorry about that, I just added you as a collaborator on the joss-reviews repo, and if you accept you should be able to check the boxes. Not sure why those permissions didn't come when I added you as a reviewer.
