### 2019-07-18 #1532 @FaustinCarter
> @FaustinCarter sorry about that, I just added you as a collaborator on the joss-reviews repo, and if you accept you should be able to check the boxes. Not sure why those permissions didn't come when I added you as a reviewer.

So I clicked on the invitation and got the following: "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." Strange. I had the privs at one point (I remember accepting the original invitation).
