### 2019-07-18 #1532 @arfon
Hi @FaustinCarter, apologies for causing you trouble here. It turns out there was a bug in our invite codes which didn't account for GitHub usernames with mixed cases (like yours). I've just fixed this an re-invited you to this repository (https://github.com/openjournals/joss-reviews/invitations).

My apologies again for any time wasted here but I believe this should now be fixed :-)
