### 2019-07-18 #1532 @dap-biospec
@FaustinCarter  Thanks for recommendations. The purpose of two versions is to keep work space clean for those who do not need development, but the point is taken. I will look in to moving internal code into an independent module, as we have done for G3F (https://github.com/dap-biospec/G3F). There, we have a single main code file while and all support  code is loaded via includes and  is hidden unless the user( developer) enables independent module development.
