### 2019-07-18 #1532 @FaustinCarter
@dap-biospec Sounds good. I'm generally open to anything that doesn't require moving files around after the repository is cloned/downloaded.
