### 2019-07-21 04d6731
Update KES_API.md
