+++ KES_API.md
@@ -9,15 +9,15 @@ Note to users: The following is a list of commands automatically called by the K
 #### function `SimCtrlTable(jobW)`
 - Description: Used to open a table showing the names and contents of all parameter waves for the indicated job. Parameters can be edited directly from this table.
 - Input:
-	- `WAVE` `/T` `jobW`: name of the text job wave. Name is set by the user.
+	- `WAVE /T jobW`: name of the text job wave. Name is set by the user.
 - Return: none
 
 ### Performing simulations
 #### function `Set_MedEChem(jobW, doSingle=int)`
 - Description: Performs a simulation set or single simulation. Equivalent to pressing the `do single set` or `do this sim` buttons on the KinESim `Control Panel`.
 - Input:
-	- `Wave` `/T` `jobW`: name of the text job wave. Name is set by the user.
-	- `Variable` `doSingle=int`: Determines if the program should perform a single simulation (`do this sim`) or a simulation set (`do single set`). Int` = `1` (true) or `Int` = `0` (false).
+	- `Wave /T jobW`: name of the text job wave. Name is set by the user.
+	- `Variable doSingle=int`: Determines if the program should perform a single simulation (`do this sim`) or a simulation set (`do single set`). Int` = `1` (true) or `Int` = `0` (false).
 	- Example: `Set_MedEchem(NPVJob, doSingle=1)`
 		- This command would perform a single simulation (`doSingle=1`) using the parameters given by the `NPVJob` wave.
 		- Return: none
@@ -26,7 +26,7 @@ Note to users: The following is a list of commands automatically called by the K
 #### function `Kilo_MedEChem05(jobW)`
 - Description: Performs a simulation kilo set. Equivalent to pressing the `do kilo set` button on the KinESim `Control Panel`.
 - Input:
-	- `WAVE` `/T` `jobW`: name of the text job wave. Name is set by the user.
+	- `WAVE /T jobW`: name of the text job wave. Name is set by the user.
 	- Example: `Kilo_MedEChem05(NPVJob)`
 - Return: none
 - History report: Output folder location. Simulation time and number of steps performed.
@@ -34,7 +34,7 @@ Note to users: The following is a list of commands automatically called by the K
 #### function `Mega_MedEChem05(jobW)`
 - Description: Performs a simulation mega set. Equivalent to pressing the `do mega set` button on the KinESim `Control Panel`.
 - Input:
-	- `WAVE` `/T` `jobW`: name of the text job wave. Name is set by the user.
+	- `WAVE /T jobW`: name of the text job wave. Name is set by the user.
 	- Example: `Mega_MedEChem05(NPVJob)`
 - Return: none
 - History report: Output folder location. Simulation time and number of steps performed.
@@ -46,74 +46,74 @@ Note to users: The following is a list of commands automatically called by the K
 #### Threadsafe function `simWSetupProto(SetData, SimData)`
 - Description: prepare method waveform
 - Input:
-	- `STRUCT` `simSetDataT` `SetData`: data structure for a simulation set.
-	- `STRUCT` `simDataT` `SetData`: data structure for a single simulation.
+	- `STRUCT simSetDataT SetData`: data structure for a simulation set.
+	- `STRUCT simDataT SetData`: data structure for a single simulation.
 - Return: none
 
 #### Threadsafe function `simWProcessProto(ResultWNSuffix, SimData)`
 - Description: prepare method waveform
 - Input:
 	- `ResultWNSuffix`: name suffix for processed waves; based off simulation set name.
-	- `STRUCT` `simDataT` `SetData`: data structure for a single simulation.
+	- `STRUCT simDataT SetData`: data structure for a single simulation.
 - Return: `WAVE` IWave: processed data
 
[…]
