### 2019-07-21 #1532 @kyleniemeyer
Hi @usethedata and @cfneese, could you submit your reviews within the next few days, or by the end of this week? Thanks!
