### 2019-07-21 #1532 @dap-biospec
Hi @usethedata and  @cfneese,  We are finishing reorganization suggested by  @FaustinCarter and will be uploading documentation and the restructured code on Monday am. This does not change functionality.

CC: kyleniemeyer
