### 2019-07-21 0ee0ce5
Update KES_API.md
