+++ KES_API.md
@@ -3,21 +3,55 @@ Return to the [readme](https://github.com/dap-biospec/KinESim/blob/master/readme
 See the [manual](https://github.com/dap-biospec/KinESim/blob/master/KES_manual.md)
 
 ## API
-Note to users: The following is a list of commands automatically called by the Kin-E-Sim `Control Panel`. Knowledge of these commands is not required to use KinESim. Full functionality is provided by the `Control Panel` however these commands can be performed manually by calling them in the Igor command line.
+
+### Contents
+
+#### [KES_Core](https://github.com/dap-biospec/KinESim/blob/master/KES_API.md#KES_Core)
+
+#### [KES_Sets](https://github.com/dap-biospec/KinESim/blob/master/KES_API.md#KES_Sets)
+
+#### [KES_GUI](https://github.com/dap-biospec/KinESim/blob/master/KES_API.md#KES_GUI)
+
+### KES_Core
+#### threadsafe function `Sim_Core_Seq(SWave, CWave, PWave, ERxnsW, GRxnsW, RxnsTmp, OWave, idx, LogWave, logMode)
+- Description: 
+- Input:
+	- `WAVE /T SWave`: 
+	- `WAVE /T CWave`: name of the components wave
+	- `WAVE /T PWave`: name of the parameters wave
+	- `WAVE /WAVE GRxnsW`: name of the GRxns wave
+	- `WAVE /WAVE ERxnsW`: name of the ERxns wave
+	- `WAVE RxnsTmp`: 
+	- `WAVE OWave`:
+	- `variable idx`: The row in which data is reported. Default start value is 0. 
+	- `WAVE LogWave`:
+	- `variable logMode`: 
+- Output: None
+
+#### function `Sim_Core_Prl(simData, OWave, idx, nMT, hostPrg)
+- Description:
+- Input:
+	- `STRUCT simDataT simData`:
+	- `WAVE OWave`:
+	- `variable idx`: The row in which data is reported. Default start value is 0.
+	- `variable nMT`: 
+	- `STRUCT SetProgDataT hostPRG`: 
+- Output: None
+
 
 ### Editing parameters
 #### function `SimCtrlTable(jobW)`
 - Description: Used to open a table showing the names and contents of all parameter waves for the indicated job. Parameters can be edited directly from this table.
 - Input:
-	- `WAVE` `/T` `jobW`: name of the text job wave. Name is set by the user.
+	- `WAVE /T jobW`: name of the text job wave. Name is set by the user.
 - Return: none
 
-### Performing simulations
+### KES_Sets
 #### function `Set_MedEChem(jobW, doSingle=int)`
 - Description: Performs a simulation set or single simulation. Equivalent to pressing the `do single set` or `do this sim` buttons on the KinESim `Control Panel`.
 - Input:
-	- `Wave` `/T` `jobW`: name of the text job wave. Name is set by the user.
-	- `Variable` `doSingle=int`: Determines if the program should perform a single simulation (`do this sim`) or a simulation set (`do single set`). Int` = `1` (true) or `Int` = `0` (false).
+	- `Wave /T jobW`: name of the text job wave. Name is set by the user.
+	- `Variable doSingle=int`: Determines if the program should perform a single simulation (`do this sim`) or a simulation set (`do single set`). Int` = `1` (true) or `Int` = `0` (false).
 	- Example: `Set_MedEchem(NPVJob, doSingle=1)`
 		- This command would perform a single simulation (`doSingle=1`) using the parameters given by the `NPVJob` wave.
 		- Return: none
@@ -26,7 +60,7 @@ Note to users: The following is a list of commands automatically called by the K
 #### function `Kilo_MedEChem05(jobW)`
[…]
