### 2019-07-22 bb8b048
restructuring repo
Repository restructured for better transparency;
Documentation updated to reflect changes in code units/file structure (v1.7.0a).
