+++ Docs/KES_API.md
@@ -1,0 +1,317 @@
+Return to the [readme](https://github.com/dap-biospec/KinESim/blob/master/readme.md)
+
+See the [manual](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md)
+## Structure
+The main procedure file (KineSim.ipf) includes several dependencies on separate untis:
+- Core: the main simulation engine. It is implemented as a independent module using namespace KES. Includes:
+	- KES_Core.ipf
+		- KES_Core_Aliases.ipf
+		- KES_Core_Flow.ipf
+		- KES_Core_Integrator.ipf
+		- KES_Core_Log.ipf
+		- KES_Core_Struct.ipf
+	
+- Sets: handles preparation and processing of individual simulations as well as one-, two-, and three-dimensional sets of simulations. It is implemented as a named module KES_Sets. Sets hands simulations off to KES core, which is required. Includes:
+	- KES_Sets.ipf
+		- KES_Sets_Aux.ipf
+		- KES_Sets_Flow.ipf
+		- KES_Sets_Progress.ipf
+		- KES_Sets_Proxy.ipf
+		- KES_Sets_Struct.ipf
+		
+- GUI: implements user interface for preparation of simulations. It is implemented as a named module KES_GUI. GUI executes functions in Sets, which reqiured. Includes:
+	- KES_GUI.ipf
+		- KES_GUI_Aux.ipf
+		- KES_GUI_Comp.ipf
+		- KES_GUI_Rxns.ipf
+		
+Upon loading, only KineSim.ipf is visible in the list of procedure files to keep user workspace clean. Visibility of all loaded units can be enabled using "SetIgorOption IndependentModuleDev=1" command or from the information panel in GUI as described in the [manual](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md).
+
+## API
+
+### Contents
+
+#### [KES_Core](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md#KES_Core)
+
+#### [KES_Sets](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md#KES_Sets)
+
+#### [KES_GUI](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md#KES_GUI)
+
+#### [Data Structures](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md#Data_Structures)
+
+### KES_Core
+#### threadsafe function `Sim_Core_Seq(SWave, CWave, PWave, ERxnsW, GRxnsW, RxnsTmp, OWave, idx, LogWave, logMode)
+- Description: 
+- Input:
+	- `WAVE /T SWave`: 
+	- `WAVE /T CWave`: name of the components wave
+	- `WAVE /T PWave`: name of the parameters wave
+	- `WAVE /WAVE GRxnsW`: name of the GRxns wave
+	- `WAVE /WAVE ERxnsW`: name of the ERxns wave
+	- `WAVE RxnsTmp`: 
+	- `WAVE OWave`:
+	- `variable idx`: The row in which data is reported. Default start value is 0. 
+	- `WAVE LogWave`:
+	- `variable logMode`: 
+- Output: None
+
+#### function `Sim_Core_Prl(simData, OWave, idx, nMT, hostPrg)
+- Description:
+- Input:
+	- `STRUCT simDataT simData`:
+	- `WAVE OWave`:
+	- `variable idx`: The row in which data is reported. Default start value is 0.
+	- `variable nMT`: 
+	- `STRUCT SetProgDataT hostPRG`: 
+- Output: None
+
+
+
+### KES_Sets
+#### function `Set_MedEChem(jobW, doSingle=int)`
[…]
