+++ readme.md
@@ -1,16 +1,19 @@
 # KinESim
 Igor Pro package and examples for simulation of homogeneous and heterogeneous electrochemical reactions
 
-**Important note to the user:** This software requires Igor v. 8.0 to run.
+**Important:** This software requires Igor v. 8.0 to run.
 
-Kin-E-Sim is a package for Igor Pro designed to predict the characteristic response of an analyte in various electrochemical methods, including normal pulsed voltammetry (illustrated below), cyclic voltammetry, etc. This document is intended to provide a new user the information they need to begin using the Kin-E-Sim program. See [citation] for further details on the implemented algorithms. 
+Kin-E-Sim is a package for Igor Pro designed to predict the characteristic response of an analyte in various electrochemical methods, including normal pulsed voltammetry (illustrated below), cyclic voltammetry, etc. This document is intended to provide a new user the information they need to begin using the Kin-E-Sim program. See [citation](https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859) for further details on the implemented algorithms. 
 
-Two versions of Kin-E-Sim are provided: `KES__complete.ipf` includes complete code in a single procedure file and is intended for end users.  `KES__Main.ipf` loads the same code which is split over several procedure files. It is intended for programmers who want to examine and revise the code.  Only one version can be loaded at a time. 
+## Installation
+Clone the git repository and open the KES/KinESim.ipf in IgorPro.  The necessary dependencies are included and are loaded by KinESim.ipf. See [API](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md#Structure) for more details of file structure. 
 
 ## Using KinESim
-A [manual](https://github.com/dap-biospec/KinESim/blob/master/KES_manual.md) is provided that includes instructions for [installation](https://github.com/dap-biospec/KinESim/blob/master/KES_manual.md#Installation), a tutorial featuring [demo simulations](https://github.com/dap-biospec/KinESim/blob/master/KES_manual.md#Kin-E-Sim-Demo), and a complete description of the KinESim [control panel](https://github.com/dap-biospec/KinESim/blob/master/KES_manual.md#Control-Panel).
+A [manual](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md) is provided that includes instructions for [installation](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md#Installation), a tutorial featuring [demo simulations](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md#Kin-E-Sim-Demo), and a complete description of the KinESim [control panel](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md#Control-Panel).
 
-An API can be found [here](https://github.com/dap-biospec/KinESim/blob/master/KES_API.md). 
+A demo of Kin-E-Sim containing pre-loaded simulations is available [here](https://github.com/dap-biospec/KinESim/blob/master/Demo).
+
+An API can be found [here](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md). 
 
 ## Community Guidelines
 Bug reports and requests for improvements, and new features are welcomed! Please feel free to make a post to [Biospec Github Issue Tracker](https://github.com/dap-biospec/KinESim/issues) or contact Denis A. Proshlyakov, [dapro@chemistry.msu.edu](mailto:dapro@chemistry.msu.edu)

[…]
