### 2019-07-22 690b94e
Restructuring of documentation for transparency
