### 2019-07-22 fe74202
Restructuring of documentation for transparency
