### 2019-07-22 #1532 @dap-biospec
@usethedata, @FaustinCarter, @cfneese - all corrections suggested by @FaustinCarter have been made and merged. Code has been reorganized as described in the updated API document. Simulation algorithms remain unchanged (apart from operating now as an independent module).

Documentation and Demo files are collected in corresponding folders. We did not include two of the in the main KinESim.ipf (as suggested by @FaustinCarter) because they are particular illustrations used in  the demo and not a common code.

We hope that repo is now in compliance with JOSS standards.

CC: @kyleniemeyer
