### 2019-07-22 90c9598
Update KES_API.md
