### 2019-07-23 #1532 @FaustinCarter
@dap-biospec This looks great! So onto testing:

Ideally there would be some sort of test suite that runs some pre-configured simulations and checks to make sure that the output matches some pre-verified output. In the absence of that, running the demo and verifying that it works will probably be the best compromise (I'm not going to ask you to write a million unit tests).

I've tried loading up the Demo, and following the directions on the demo readme (which I think has an  error: see the [insert image] text). When I open the demo I already have the full control panel there, so I think several of the steps in the instructions don't apply (i.e. the procedures are getting compiled on load already). So I click "do this sim" after I open the pxp, and I get an error "While executing a wave read, the following error occurred: index out of range". This is with the red arrow in data browser pointing to: "single analyte direct electrochemistry". If I drag the red arrow to a different folder and hit refresh in the KinESim window, I get a plot, but:

I don't really have any metrics for what is right, how to tell if the demo is doing what it is supposed to etc. It would be really helpful to include a step-by-step walk-through of doing a simple simulation with the demo, along with graphics showing expected outputs of the sim. Your interface is complex enough that a few helpful examples using the demo would go a long way towards building user-confidence that what they are getting output-wise is what they should get. To make sure that you have the same experience as your users, I'd download the pxp currently hosted on github and start from there.

I'm also totally lost with the procedures included in EChem and the panel that comes along for the ride under analysis. It is probably worth having separate API documentation for them, or a link to wherever that documentation exists if they were pulled in from another package. (Also, the Spectro E Chem panel doesn't seem to be closable by clicking on the red arrow, although maybe that is an Igor bug, as I'm noticing the same issue with the generated graph windows).
