### 2019-07-30 ef273b0
Demo and testing update
Added performance test;
Updated demo;
Updated documentation;
Minor bug fixes.
