+++ Demo/readme.md
@@ -1,53 +1,136 @@
 # Kin-E-Sim Demo
-A demo with pre-loaded simulations is provided in the `KES_Demo.pxp` file. A description of each pre-loaded simulation and instructions on how to perform them are provided below. Kin-E-Sim simulations are performed using the `Control Panel`. Full details of the `Control Panel` are provided in the following section. 
+The KinESim Demo file (KES_Demo.pxp) is included with the KinESim instillation. See the [manual](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md) for instillation instructions.
 
-- **Single analyte direct electrochemistry:** This simulation simulates normal pulsed voltammetry of a single analyte that undergoes direct electrochemistry on the electrode. As loaded, the analyte has a redox potential of -0.1 V. 
-		 
-- **Single analyte mediated electrochemistry:** This simulates NPV of a single analyte and mediator. The analyte is unable to perform any electron transfer on the electrode and is completely dependent on the mediator for redox transitions. As loaded, both the analyte and mediator have a redox potential of -0.1 V. The concentration of the analyte is 1 mM and the concentration of the mediator is 200 ÂµM. 
-- **Single analyte mediator with two redox transitions:** This simulation simulates NPV of a single analyte and mediator where the mediator has two distinct redox transitions. As loaded, the analyte has a redox potential of -0.1 V and the mediator has two redox potentials at 0.0 V and -0.2 V. The concentration of the analyte is 1 mM and concentration of the mediator is 200 ÂµM total. 
->**Hint:** A single mediator with two distinct redox transitions is accomplished by making two separate components and connecting them using the `alias this` function. Each component must start with a concentration value greater than 0 and the sum of the two concentrations is the total concentration of the mediator.
+## Performing a simulation
+Open KES_Demo.pxp file. 
 
-## Installation:
-- Clone the git repository and open the KES/KinESim.ipf in IgorPro.  The necessary dependencies are included and are loaded by KinESim.ipf. See [API](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md#Structure) for more details of file structure. 
-	- Open Igor Pro. Select `File` -> `Open File` -> `Procedureâ€¦`
-	- Locate the folder containing the necessary files. Select `KinESim.ipf` and `Open`. 
-	- A window containing the procedure will appear within the Igor workspace. Close the window. A dialog box will appear. Select `Hide`. This will compile the procedure and hide the procedure from view. Selecting `Kill` will permenantly remove the procedure from the current Igor experiment.
-	- Repeat this step for `NPV_Illustrations.ipf` and `Spectroechem.ipf`.  
-- Once Kin-E-Sim is loaded and compiled, it can be accessed by selecting `Analysis` -> `Kin-E-Sim` -> `Control Panel`. The `Control Panel` will appear.  
-- Finally, select `Analysis` -> `Kin-E-Sim` -> `Create Set`. A dialog box will appear asking for a name. Any name can be entered here. Click `Continue`. All necessary waves will be generated.
-- From the `Control Panel` select the job that you just named to load the waves into the `Control Panel`.
+To open the KinESim control panel click `Analysis -> Kin-E-Sim -> Control Panel`
 
-## Performing a single simulation
+A simulation is already set up, it just needs to be loaded into the panel. 
 
-- First, use the `Data Browser` (`Data` -> `Data Browser`) in Igor to select the folder with the title of the simulation you would like to run. This is done by clicking and dragging the red arrow to the desired experiment folder. Then click **[insert image]** in the top-left corner of the Kin-E-Sim `Control Panel`. The simulation should now be loaded containing all the pre-set commands and parameters.
+In the top-left corner of the panel, click the drop down menu labeled `job` and select `NPVJob`.
+
[…]
