### 2019-07-30 1b993be
Update KES_manual.md
