+++ Demo/readme.md
@@ -36,7 +36,7 @@ Click `do this sim` and the graph should change to look like this:
 
 This gives us a much better view of how our analyte behaves. 
 
-Now letís try simulating a scenario.
+Now let‚Äôs try simulating a scenario.
 
 In this scenario, we have a single analyte with slow electrochemical kinetics on the electrode. 
 
@@ -53,18 +53,18 @@ Notice that our analyte is no longer being completely reduced or oxidized at eac
 Our goal now is to find a mediator that will improve the rate of electron transfer to and from the analyte so that we can achieve a profile like we saw in Figure 2.
 
 ## Adding Components
-Letís add a new component that will represent our mediator to our components panel.
+Let‚Äôs add a new component that will represent our mediator to our components panel.
 To do this click the `+C#` button in the `components` panel
 
-A new component has been made but it needs a name. Type ìMediator 1î into the text box.
+A new component has been made but it needs a name. Type ‚ÄúMediator 1‚Äù into the text box.
 
 You can switch between the analyte and mediator 1 by clicking the drop down menu labeled `C #` and selecting `0` (the analyte) or `1` (the mediator). 
 
-Letís give our mediator some electrochemical properties. The default value for the redox potential (`E0`) is 0 V. Keep this value 0. Change `n` to 1. This means that the mediator transfers 1 electron per mediator molecule. Change `alpha` to 0.5. This represents the alpha of the mediator. This value is typically assumed to be 0.5 for most electrochemically active species.
+Let‚Äôs give our mediator some electrochemical properties. The default value for the redox potential (`E0`) is 0 V. Keep this value 0. Change `n` to 1. This means that the mediator transfers 1 electron per mediator molecule. Change `alpha` to 0.5. This represents the alpha of the mediator. This value is typically assumed to be 0.5 for most electrochemically active species.
 
 We want a mediator with fast kinetics so change `ET rate` to 0.2 and `lim. Rate` to 0.5.
 
-Now that weíve described our mediator we still need to add some amount of the mediator to our simulated electrochemical reaction. Since we are performing oxidation, we will start with a completely reduced sample. `[ox.]` is the concentration of oxidized mediator in the sample in molarity and should remain 0. Change `[rd.]`, the concentration of reduced mediator in the sample to 0.0002 M. 
+Now that we‚Äôve described our mediator we still need to add some amount of the mediator to our simulated electrochemical reaction. Since we are performing oxidation, we will start with a completely reduced sample. `[ox.]` is the concentration of oxidized mediator in the sample in molarity and should remain 0. Change `[rd.]`, the concentration of reduced mediator in the sample to 0.0002 M. 
 
 Click `do this sim` and see how it looks:
 
@@ -78,7 +78,7 @@ To fix this we need to add an E-chem reaction.
 
 To do this, click the ![alt text](https://github.com/dap-biospec/KinESim/blob/master/Docs/Figures/Rxns_create.png) button in the `E-Chem reactions` panel. This will generate the Igor waves needed to describe an E-Chem reaction. 
 
-Name the reaction ìM1+Aî by typing the name in the text box of the `E-chem reactions` panel. 
+Name the reaction ‚ÄúM1+A‚Äù by typing the name in the text box of the `E-chem reactions` panel. 
 
 Below is a table that describes the E-chem reaction. See the [manual](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_manual.md#Echem-reactions) for a description of the table. 
 
@@ -92,7 +92,7 @@ We need to add a row so that there is one row for the analyte and one row for th
 
 Click the `+ row` button to add a row. 
 
-Now fill in the table as shown below:
+Now fill in the table as shown below pressing the enter or return key after typing each number:
 
[…]
