### 2019-07-30 #1532 @dap-biospec
@usethedata, @FaustinCarter, @cfneese  we addressed all the issues that were raised by @FaustinCarter.
- We overhauled the demo and associated .md description. Demo now takes the user step by step through two simulation examples.
- The EChem unit is no longer included - most of the code was not related to KinESim and served more  specialized purpose that would  justify its inclusion here. The remaining code in NPV_Illustrations.ipf  is only an illustration of how user-supplied functions can be used to set up and process the results. Common templates are described in the "Prototype functions ..." section of the API.
- Minor bugs that caused the reported error were fixed.
- We added a separate performance test experiment. It contains a pre-configured simulation and several graphs that provide visual reference of expected simulation results under various conditions.

CC: @kyleniemeyer
