### 2019-08-04 #1532 @usethedata
Thanks.  I've done some basic work with this today, and I'll go through and finish my review in the next couple of days.  Thanks, to @FaustinCarter for your work in thinking through constructive suggestions for the authors.
