### 2019-08-05 #1532 @FaustinCarter
I think I'm about ready to recommend for publication on my end. The image links in the updated markdowns are broken though. I'd suggest using relative image links in the files so that you don't have to update all the image links every time a new commit/version happens.

Also, in the paper, the references should list a DOI (or equivalent for arxiv).

Once those are fixed, the version number is bumped, and a new release is generated, I'll tick off the final boxes.
