### 2019-08-05 9f1155d
Update KES_manual.md
Updated image links
