+++ Docs/KES_manual.md
@@ -1,6 +1,6 @@
-﻿Return to the [readme](https://github.com/dap-biospec/UTEIU1dq7RRTfvV9/blob/master/test/readme.md)
+﻿Return to the [readme](https://github.com/dap-biospec/KinESim/blob/master/readme.md)
 
-See the [API](https://github.com/dap-biospec/UTEIU1dq7RRTfvV9/blob/master/test/Docs/KES_API.md)
+See the [API](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md)
 
 ## Installation:
 - Clone the git repository and open the KES/KinESim.ipf in IgorPro.  The necessary dependencies are included and are loaded by KinESim.ipf. See [API](https://github.com/dap-biospec/KinESim/blob/master/Docs/KES_API.md#Structure) for more details of file structure. 
@@ -39,12 +39,12 @@ A demo of Kin-E-Sim containing pre-loaded simulations is available [here](https:
 A performance test is available [here](https://github.com/dap-biospec/KinESim/tree/master/Demo/PerformanceTest.md).
 
 ## Control Panel
-![alt text](https://github.com/dap-biospec/UTEIU1dq7RRTfvV9/blob/master/test/Docs/Figures/ControlPanel.png)
+![alt text](https://github.com/dap-biospec/KinESim/blob/master/Docs/Figures/ControlPanel.png)
 The `Control Panel` used to perform Kin-E-Sim simulations is shown above. Each sub-panel and its functions are described below:
 ### Top-left corner
-![alt text](https://github.com/dap-biospec/UTEIU1dq7RRTfvV9/blob/master/test/Docs/Figures/TLC.png)
-- ![alt text](https://github.com/dap-biospec/UTEIU1dq7RRTfvV9/blob/master/test/docs/Figures/TLC_i.png): Shows the current version of Kin-E-Sim kinetic simulator
-- ![alt text](https://github.com/dap-biospec/UTEIU1dq7RRTfvV9/blob/master/test/docs/Figures/TLC_refresh.png): Reloads selected job
+![alt text](https://github.com/dap-biospec/KinESim/blob/master/Docs/Figures/TLC.png)
+- ![alt text](https://github.com/dap-biospec/KinESim/blob/master/Docs/Figures/TLC_i.png): Shows the current version of Kin-E-Sim kinetic simulator
+- ![alt text](https://github.com/dap-biospec/KinESim/blob/master/Docs/Figures/TLC_refresh.png): Reloads selected job
 - `job`: Allows for the selection of a user-defined job set that specifies all values, methods, and waves for the simulation. Job sets are stored in text waves. Multiple job waves can be made to easily switch between experiments. 
 - `sets=>`: Defines parameters of single simulation or simulation sets.
 - `ctrl. table`: Creates a table showing the names and contents of all parameter waves for the selected job. Parameters can be edited from this table. Press   to update changes in the Control Panel.
@@ -52,7 +52,7 @@ The `Control Panel` used to perform Kin-E-Sim simulations is shown above. Each s
 - `flags`: miscellaneous flags
 - `esim =>`: Defines the configuration of the electrochemical system and parameters of the numerical integrator. 
 ### Simulation
-![alt text](https://github.com/dap-biospec/UTEIU1dq7RRTfvV9/blob/master/test/docs/Figures/simulation_sp.png)
+![alt text](https://github.com/dap-biospec/KinESim/blob/master/Docs/Figures/simulation_sp.png)
 - `prep`: function that generates the potential waveform or performs other pre-simulation steps
 - `process`: function that performs step-wise simulation and integration
[…]
