### 2019-08-05 fb63001
DOI updated
Updated known DOI for references.
