+++ paper.bib
@@ -3,7 +3,8 @@
 	Journal = {Analytical Chemistry},
 	Title = {{FTIR Spectro-Voltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures}},
 	Year = 2019,
-	adsurl = {https://pubs.acs.org/doi/full/10.1021/acs.analchem.9b00859}
+	doi = {10.1021/acs.analchem.9b00859},
+	adsurl = {https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859}
 }
 
 @article{John:2019b,
@@ -19,7 +20,9 @@
 	Booktitle = {Design and Analysis of Biomolecular Circuits: Engineering Approaches to Systems and Synthetic Biology, by H. Koeppl and H. Densmore and G. Setti and M. di Bernardo.~ISBN 978-1-4419-6766-4.~Published by Springer, 2011.},
 	Publisher = {Springer},
 	Title = {{Continuous Time Markov Chain Models for Chemical Reaction Networks}},	
-	Year = 2011
+	Year = 2011,
+	adsurl = {https://doi.org/10.1007/978-1-4419-6766-4_1},
+	doi = {10.1007/978-1-4419-6766-4_1}
 }
 
 @article{Zhang:2018,
@@ -29,6 +32,7 @@
 	Month = feb,
 	Title = {{Molecular Computing for Markov Chains}},
 	Year = 2018,
-	adsurl = {https://arxiv.org/abs/1802.05170}
+	adsurl = {https://arxiv.org/abs/1802.05170},
+	doi = {10.1007/s11047-019-09736-8}
 }	
 
