### 2019-08-05 #1532 @dap-biospec
@FaustinCarter, @usethedata  - links have been fixed. Apologies - while we are learning, we use testing repo to verify content and some replacements slipped through. Relative links is a good idea.

Known DOI have been updated, One of our cited papers will be re-submitted with minor requested revisions tomorrow (8/6/19) and it does not have a DOI yet. The editor did not respond to me regarding posting it in a pre-publication archive, I will follow up on it.

I do not see a way to update the version. Is this something that we can do?

I second big thanks to @FaustinCarter for patience and guidance!
