### 2019-08-05 #1532 @FaustinCarter
Regarding versioning:

It's good practice to put your version somewhere in your code base. Since IGOR doesn't have a well-defined way to do this, two possible (out of many) options are:

 - Use a VERSION.md file
 - Make a wave in your top level .ipf file called "version" and hardcode the number into it

In either case, after you do that and commit, you'll want to add a git tag with the new version number and then draft a new release on github.

It is also good practice to include a CHANGELOG.md file in your top level repo directory that enumerates changes, updates, breaking changes, etc. I'm not going to hold the review up over this though.

I've ticked all the boxes except the version one. You may want to wait till the other reviewers have chimed in with any changes they may want before finalizing the version. In any case, I am ready to recommend for publication.
