### 2019-08-06 #1532 @dap-biospec
Thank you, @FaustinCarter.  We have version coded in KinESim as a string constants in the root procedure files. These versions are displayed in GUI "info" dialog and core version is printed in the history (the latest version is 1.7.0a).  Dependencies are checked for the minimal file date stamp (implemented via "#pragma version = yyyymmdd" integer value) and Igor will alert the user if any dependent files are out of date.

The last major GitHub commit has version info displayed in the summary and the list of major changes is in the commit details. I agree that we need to set up changelog.md with the future updates.

What I do not know is how to update the version initially declared with JOSS submission (v1.6.8g) to the latest version that you see (1.7.0a). I do not see this information in the paper.md or paper.bib documents and it seems to be beyond my reach.

CC: @kyleniemeyer
