### 2019-08-06 #1532 @kyleniemeyer
I can change the version here—is v1.7.0a is the final version?
