### 2019-08-06 #1532 @dap-biospec
@kyleniemeyer this is correct, v1.7.0a is the final version unless further changes are requested.
