### 2019-08-10 df393f7
Added ChemRxiv reference
