+++ paper.bib
@@ -8,11 +8,12 @@
 }
 
 @article{John:2019b,
-	Adsnote = {Placeholder for paper under intial review in The Journal of Physical Chemistry B},
+	Adsnote = {submitted to The Journal of Physical Chemistry B},
 	Author = {{John}, C.~W., and {Swain}, G.~M. and {Hausinger}, R.~P. and {Proshlyakov}, D.A.},
-	Journal = {The Journal of Physical Chemistry B},
+	Journal = {ChemRxiv e-prints},
 	Title = {{Strongly coupled redox-linked conformational switching at the active site of the non-heme iron-dependent dioxygenase, TauD}},
-	Year = 2019
+	Year = 2019,
+	doi = {10.26434/chemrxiv.9461462}
 }
 
 @book{Anderson:2011,
