### 2019-08-10 #1532 @usethedata
Sorry my review took a lot longer than it should, but it is complete.  I very much appreciate the work that @FaustinCarter did.  There's nothing I see that needs to be changed and I recommend publication.
