### 2019-08-10 #1532 @dap-biospec
@kyleniemeyer we received permission from the editor at J. Phys. Chem. to deposit preprint to ChemRxiv. Bibliography now includes DOI for all cited items. We hope that this will allow to complete the review.

cc: @cfneese
