### 2019-08-23 2aa9bf3
Reference for an accepted article updated
