+++ paper.bib
@@ -1,39 +1,43 @@
 @article{John:2019a,
-	Author = {{John}, C.~W. and {Proshlyakov}, D.~A.},
-	Journal = {Analytical Chemistry},
-	Title = {{FTIR Spectro-Voltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures}},
-	Year = 2019,
-	doi = {10.1021/acs.analchem.9b00859},
-	adsurl = {https://pubs.acs.org/doi/10.1021/acs.analchem.9b00859}
+  doi = {10.1021/acs.analchem.9b00859},
+  year = 2019,
+  publisher = {American Chemical Society ({ACS})},
+  volume = {91},
+  number = {15},
+  pages = {9563--9570},
+  author = {Christopher W. John and Denis A. Proshlyakov},
+  title = {Fourier Transform Infrared Spectrovoltammetry and Quantitative Modeling of Analytes in Kinetically Constrained Redox Mixtures},
+  journal = {Analytical Chemistry}
 }
 
 @article{John:2019b,
-	Author = {{John}, C.~W., and {Swain}, G.~M. and {Hausinger}, R.~P. and {Proshlyakov}, D.A.},
-	Journal = {Journal of Physical Chemistry B},
-	Title = {{Strongly coupled redox-linked conformational switching at the active site of the non-heme iron-dependent dioxygenase, TauD}},
-	Year = 2019,
-	doi = {10.1021/acs.jpcb.9b05866},
-	adsurl = {https://pubs.acs.org/doi/10.1021/acs.jpcb.9b05866}
+  doi = {10.1021/acs.jpcb.9b05866},
+  year = 2019,
+  month = aug,
+  publisher = {American Chemical Society ({ACS})},
+  author = {Christopher W. John and Greg M Swain and Robert P. Hausinger and Denis A. Proshlyakov},
+  title = {Strongly Coupled Redox-Linked Conformational Switching at the Active Site of the Non-Heme Iron-Dependent Dioxygenase, {TauD}},
+  journal = {The Journal of Physical Chemistry B}
 }
 
-@book{Anderson:2011,
-	Author = {{Anderson}, D. and {Kurtz}, T.},
-	Booktitle = {Design and Analysis of Biomolecular Circuits: Engineering Approaches to Systems and Synthetic Biology, by H. Koeppl and H. Densmore and G. Setti and M. di Bernardo.~ISBN 978-1-4419-6766-4.~Published by Springer, 2011.},
-	Publisher = {Springer},
-	Title = {{Continuous Time Markov Chain Models for Chemical Reaction Networks}},	
-	Year = 2011,
-	adsurl = {https://doi.org/10.1007/978-1-4419-6766-4_1},
-	doi = {10.1007/978-1-4419-6766-4_1}
+@book{Anderson:2011,	
+  Title = {{Continuous Time Markov Chain Models for Chemical Reaction Networks}},	
+  doi = {10.1007/978-1-4419-6766-4_1},
+  year = 2011,
+  publisher = {Springer New York},
+  pages = {3--42},
+  author = {David F. Anderson and Thomas G. Kurtz},
+  booktitle = {Design and Analysis of Biomolecular Circuits},
+  editor = {H. Koeppl and G. Setti and M. {di Bernardo} and D. Densmore},
 }
 
 @article{Zhang:2018,
-	Author = {{Zhang}, C. and {Shen}, Z. and {Wei}, W. and {Zhao}, J. and {Zhang}, Z., and {You}, X.},
-	Eprint = {1802.05170},
-	Journal = {ArXiv e-prints},
-	Month = feb,
[…]
