### 2019-08-23 ecab8b3
Update paper.md
