+++ paper.md
@@ -27,11 +27,13 @@ bibliography: paper.bib
 
 ## Rationale:
 Electrochemistry studies reduction-oxidation reactions, or reactions that involve transfer of an electron, with a particular emphasis on the reactions occurring at the interface between the electrode and the solution. Multiple electrochemical techniques can provide valuable insight into the thermodynamic properties of analytes and their chemical mechanisms. Extensive efforts have been invested into developing a quantitative description of the dynamics of electron transfer and ensuing mass transport where the analyte itself can undergo direct reduction/oxidation on the surface of the electrode. However, more complex cases where a redox-active analyte requires one or more redox-active mediators to transfer an electron to/from the electrode received less attention. Additional complexity may arise from the kinetic limitations of such multi-step reactions and the inherent nature of electrochemistry: it relies on the detection of electric currents, which report on the rate of a chemical process, while often focusing on the thermodynamic properties of the analyte, which represents equilibrium conditions. This conundrum requires a balance between detectability (fast processes) and accuracy (slow processes). Therefore, in most cases electrochemical studies are conducted under pre-equilibrium conditions.
-KinESim is a procedure package for Igor Pro that was primarily designed to predict pre-equilibrium concentrations and their changes in a multi-component, redox-active chemical mixture. In general, this includes homogenous electrochemical reactions in the solution and heterogenous processes on the solution/electrode boundary. Heterogenous electrochemical reactions are controlled by an applied electric potential of an arbitrary waveform (continuous or pulsed). This allows one to use KinESim to predict pre-equilibrium changes in a variety of standard techniques, such as pulse, staircase, or cyclic voltammetry, as well as custom potential profiles. KinESim supports branching, sequential, and isomerization reactions in various stochiometric ratios over both phases. Unlike experimental electrochemical methods that report net electric current/charge, KinESim provides an insight into changes occurring to the individual reagents and their redox sub-population. This is particularly significant for spectro-electrochemical and indirect (mediated) electrochemical studies where redox changes in the analytes of interests may be difficult to detect from the electric current alone. Therefore, it can be used to predict experimental data and to test proposed reaction mechanisms in most chemical processes controlled by a combination of kinetics and thermodynamics. These abilities have recently been demonstrated in a study of how electrochemical mediators affect the observed electrochemical properties of an analyte [@John:2019a;@John:2019b].
+
+KinESim is a procedure package for Igor Pro that was primarily designed to predict pre-equilibrium concentrations and their changes in a multi-component, redox-active chemical mixture. In general, this includes homogenous electrochemical reactions in the solution and heterogenous processes on the solution/electrode boundary. Heterogenous electrochemical reactions are controlled by an applied electric potential of an arbitrary waveform (continuous or pulsed). This allows one to use KinESim to predict pre-equilibrium changes in a variety of standard techniques, such as pulse, staircase, or cyclic voltammetry, as well as custom potential profiles. KinESim supports branching, sequential, and isomerization reactions in various stochiometric ratios over both phases. Unlike experimental electrochemical methods that report net electric current/charge, KinESim provides an insight into changes occurring to the individual reagents and their redox sub-population. This is particularly significant for spectro-electrochemical and indirect (mediated) electrochemical studies where redox changes in the analytes of interests may be difficult to detect from the electric current alone. Therefore, it can be used to predict experimental data and to test proposed reaction mechanisms in most chemical processes controlled by a combination of kinetics and thermodynamics. These abilities have recently been demonstrated in a study of how electrochemical mediators affect the observed electrochemical properties of an analyte [@John:2019a; @John:2019b].
 
 ## Concept:
[…]
