### 2019-08-23 #1532 @kyleniemeyer
@dap-biospec I made some minor edits to the paper, please merge these PRs: https://github.com/dap-biospec/KinESim/pull/5 and https://github.com/dap-biospec/KinESim/pull/6
