### 2019-08-23 #1532 @kyleniemeyer
Congrats @dap-biospec on your article's publication in JOSS! Many thanks to @usethedata and @FaustinCarter for reviewing.
